+++ tests/ParserExceptionTest.php
@@ -1,0 +1,72 @@
+<?php
+
+declare(strict_types=1);
+
+/**
+ * Tests that parsers signal invalid content with ParserException.
+ */
+
+namespace SugiPHP\Config\Tests;
+
+use SugiPHP\Config\Exception\ParserException;
+use SugiPHP\Config\Parser\Json;
+use SugiPHP\Config\Parser\Php;
+use SugiPHP\Config\Parser\Xml;
+use PHPUnit\Framework\TestCase;
+
+class ParserExceptionTest extends TestCase
+{
+    public function testMalformedXmlThrowsParserException()
+    {
+        $this->expectException(ParserException::class);
+        (new Xml())->parse('<config><host>localhost</host>');
+    }
+
+    public function testMalformedXmlDoesNotEmitWarnings()
+    {
+        set_error_handler(function ($errno, $errstr) {
+            $this->fail("Unexpected PHP warning/notice: {$errstr}");
+        });
+
+        try {
+            (new Xml())->parse('<config>');
+            $this->fail('Expected ParserException');
+        } catch (ParserException $e) {
+            $this->assertStringContainsString('XML parse error', $e->getMessage());
+        } finally {
+            restore_error_handler();
+        }
+    }
+
+    public function testValidXmlStillParses()
+    {
+        $this->assertSame(["host" => "localhost"], (new Xml())->parse('<config><host>localhost</host></config>'));
+    }
+
+    public function testMalformedJsonThrowsParserException()
+    {
+        $this->expectException(ParserException::class);
+        (new Json())->parse('{"host": ');
+    }
+
+    public function testNonArrayJsonThrowsParserException()
+    {
+        $this->expectException(ParserException::class);
+        (new Json())->parse('42');
+    }
+
+    public function testPhpFileNotReturningArrayThrowsParserException()
+    {
+        $tmp = tempnam(sys_get_temp_dir(), 'sugi');
+        $file = $tmp . '.php';
+        file_put_contents($file, '<?php return "not an array";');
+
+        try {
+            $this->expectException(ParserException::class);
+            (new Php())->parseFile($file);
+        } finally {
+            unlink($file);
+            unlink($tmp);
+        }
+    }
+}